--- app/src/main/java/ed/maevski/androidpraktika/DetailsActivity.kt
@@ -1,25 +1,0 @@
-package ed.maevski.androidpraktika
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import ed.maevski.androidpraktika.data.DeviantPicture
-import ed.maevski.androidpraktika.databinding.ActivityDetailsBinding
-
-class DetailsActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityDetailsBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityDetailsBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val pic = intent.extras?.get("dev") as DeviantPicture
-
-        //Устанавливаем заголовок
-        binding.detailsToolbar.title = pic.title
-        //Устанавливаем картинку
-        binding.detailsPoster.setImageResource(pic.picture)
-        //Устанавливаем описание
-        binding.detailsDescription.text = pic.description
-    }
-}
