+++ app/src/main/java/ed/maevski/androidpraktika/adapter/PictureDelegateAdapter.kt
@@ -31,14 +31,11 @@ class PictureDelegateAdapter(private val clickListener: PictureRecyclerAdapter.O
         holder.author.text = item.author
         holder.description.text = item.description
 
-        //Обрабатываем нажатие на весь элемент целиком(можно сделать на отдельный элемент
+        //Обрабатываем нажатие на весь элемент целиком - Уточнить????  (можно сделать на отдельный элемент
         //например, картинку) и вызываем метод нашего листенера, который мы получаем из
         //конструктора адаптера
         holder.picture.setOnClickListener{
             clickListener.click(item)
         }
-/*            .item_container.setOnClickListener {
-                clickListener.click(items[position])
-            }*/
     }
 }
