+++ app/src/main/res/layout/item_ad.xml
@@ -2,19 +2,19 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:paddingStart="15dp"
-    android:paddingEnd="15dp"
+    android:paddingStart="@dimen/paddingsLayoutItemAd"
+    android:paddingEnd="@dimen/paddingsLayoutItemAd"
     android:orientation="vertical">
 
     <TextView
         android:id="@+id/title"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:textSize="23sp" />
+        android:textSize="@dimen/textSizeTitleItemAd" />
 
     <TextView
         android:id="@+id/content"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:textSize="18sp" />
+        android:textSize="@dimen/textSizeContentItemAd" />
 </LinearLayout>
