+++ app/src/main/res/layout/item_picture.xml
@@ -5,9 +5,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="150dp"
-    android:elevation="1dp"
-    app:cardCornerRadius="10dp">
+    android:layout_height="@dimen/heightLayoutItemPicture"
+    android:elevation="@dimen/elevationLayoutItemPicture"
+    app:cardCornerRadius="@dimen/radiusLayoutItemPicture">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -15,10 +15,10 @@
 
         <ImageView
             android:id="@+id/poster"
-            android:layout_width="150dp"
+            android:layout_width="@dimen/widthImageItemPicture"
             android:layout_height="match_parent"
             android:scaleType="centerCrop"
-            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             tools:src="@tools:sample/avatars" />
 
@@ -27,10 +27,10 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:textAlignment="center"
-            android:padding="8dp"
+            android:padding="@dimen/paddingsTextItemPicture"
             android:textAppearance="@style/TextAppearance.AppCompat.Headline"
-            app:layout_constraintLeft_toRightOf="@+id/poster"
-            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/poster"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             tools:text="Title" />
 
@@ -39,9 +39,9 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:textAlignment="center"
-            android:padding="8dp"
-            app:layout_constraintLeft_toRightOf="@+id/poster"
-            app:layout_constraintRight_toRightOf="parent"
+            android:padding="@dimen/paddingsTextItemPicture"
+            app:layout_constraintStart_toEndOf="@+id/poster"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/title"
             tools:text="Author" />
 
@@ -49,10 +49,10 @@
             android:id="@+id/description"
             android:layout_width="0dp"
             android:layout_height="0dp"
-            android:padding="8dp"
+            android:padding="@dimen/paddingsTextItemPicture"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toRightOf="@+id/poster"
-            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/poster"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/author"
             tools:text="@tools:sample/lorem/random" />
