+++ app/src/main/res/values/strings.xml
@@ -6,6 +6,5 @@
     <string name="menu_selections_title">подборки</string>
     <string name="page_title">поиск</string>
     <string name="disclaimer">content is from deviantart.com</string>
-
-
+    <string name="textToExit">Double tap for exit</string>
 </resources>
